[refactor] change package structure
