+++ backend/src/main/java/com/woowacourse/edd/presentation/exceptionhandler/Error.java
@@ -1,4 +1,4 @@
-package com.woowacourse.edd.domain.errors;
+package com.woowacourse.edd.presentation.exceptionhandler;
 
 public class Error {
     private String result;
