+++ backend/src/main/java/com/woowacourse/edd/repository/VideoRepository.java
@@ -1,5 +1,6 @@
-package com.woowacourse.edd.domain;
+package com.woowacourse.edd.repository;
 
+import com.woowacourse.edd.domain.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VideoRepository extends JpaRepository<Video, Long> {
