+++ backend/src/test/java/com/woowacourse/edd/application/service/VideoServiceTest.java
@@ -2,12 +2,12 @@ package com.woowacourse.edd.application.service;
 
 import com.woowacourse.edd.application.converter.VideoConverter;
 import com.woowacourse.edd.domain.Video;
-import com.woowacourse.edd.domain.VideoRepository;
+import com.woowacourse.edd.repository.VideoRepository;
 import com.woowacourse.edd.domain.vo.Contents;
 import com.woowacourse.edd.domain.vo.Title;
 import com.woowacourse.edd.domain.vo.YoutubeId;
-import com.woowacourse.edd.presentation.controller.VideoResponse;
-import com.woowacourse.edd.presentation.dto.VideoSaveRequestDto;
+import com.woowacourse.edd.application.response.VideoResponse;
+import com.woowacourse.edd.application.dto.VideoSaveRequestDto;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
